requirements.txt
# Multi-Agent Design System Requirements

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Image Processing
Pillow>=10.1.0

# Async Support
asyncio-mqtt>=0.16.0
aiofiles>=23.2.0

# HTTP Client
httpx>=0.25.0

# Data Validation
email-validator>=2.1.0

# Development Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Optional: Future Hugging Face Integration
# transformers>=4.35.0
# diffusers>=0.21.0
# torch>=2.1.0
# accelerate>=0.24.0

# Optional: Database Support
# sqlalchemy>=2.0.0
# alembic>=1.12.0
# asyncpg>=0.29.0

# Optional: Redis for Task Queue
# redis>=5.0.0
# celery>=5.3.0

# Optional: Monitoring
# prometheus-client>=0.19.0
# structlog>=23.2.0
